% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{har.mean}
\alias{har.mean}
\title{Harmonic mean}
\usage{
har.mean(x)
}
\arguments{
\item{x}{vector of numerical values}
}
\description{
Harmonic mean
}
\examples{
#generate some data
data <- rnorm(10)

#calculate harmonic mean
har.mean(data)

}
