% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x, nbins=NA)
}
\arguments{
\item{x}{vector of numerical or categorical values}

\item{nbins}{Number of bins to use for continuous data}
}
\description{
Mode
}
\examples{
#generate some data
data <- rnorm(10)

#calculate mode
mode(data)

}
