% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{se}
\alias{se}
\title{Standard error of the mean}
\usage{
se(x)
}
\arguments{
\item{x}{vector of numerical values}
}
\description{
Standard error of the mean
}
\examples{
#generate some data
data <- rnorm(10)

#calculate the standard error
se(data)

}
